requirements.txt
asgiref==3.2.7
Django==3.0.4
django-admin-menu==1.2.1
django-appconf==1.0.4
django-compressor==2.4
django-crispy-forms==1.9.0
django-currentuser==0.5.0
django-filter==2.2.0
django-libsass==0.8
django-material-admin==1.7.0
django-widget-tweaks==1.4.8
jsonfield==3.1.0
libsass==0.19.4
psycopg2==2.8.4
pytz==2019.3
rcssmin==1.0.6
rjsmin==1.1.0
six==1.14.0
sqlparse==0.3.1